webapp/requirements.txt
waitress==3.0.0
requests==2.31.0
flask==3.0.2
Flask-Session==0.6.0
PyYAML==6.0.1
Pillow==10.0.0
PyMySQL
cryptography
bcrypt
requests
Flask-WTF